加入unistd.h和函数sysconf

+++ include/string.h
@@ -1,14 +1,16 @@
 #ifndef _STRING_H
 #define _STRING_H
 
-void *memcpy(void *dst, const void *src, unsigned long len);
-void *memset(void *b, int c, unsigned len);
-int memcmp(const void *b1, const void *b2, unsigned long len);
+#include <stddef.h>
+
+void *memcpy(void *dst, const void *src, size_t len);
+void *memset(void *b, int c, size_t len);
+int memcmp(const void *b1, const void *b2, size_t len);
 
 char *strcat(char *s, const char *append);
 int strcmp(const char *s1, const char *s2);
 char *strcpy(char *to, const char *from);
-unsigned strlen(const char *str);
-char *strncpy(char *dest, const char *src, unsigned n);
+size_t strlen(const char *str);
+char *strncpy(char *dest, const char *src, size_t n);
 
 #endif /* _STRING_H */
